+++ crates/windows/src/tree/live_read_tests.rs
@@ -1,10 +1,14 @@
+use super::imp::{corroborate_verified_process, read_live_element_core};
 use super::{live_actions, live_bounds, live_element, live_state, live_value, read_live_element};
-use crate::tree::automation::root_from_hwnd;
+use crate::system::hresult::UIA_E_ELEMENTNOTAVAILABLE;
+use crate::tree::automation::{UiaFailure, root_from_hwnd, uia_failure_error};
 use crate::tree::element::UIAElement;
 use crate::tree::fixture::{HostedFixture, ensure_test_apartment};
+use crate::tree::properties::{ElementProperties, read_live};
 use crate::tree::walker::TreeSource;
 use crate::tree::walker_fake::deadline;
 use agent_desktop_core::{AdapterError, ErrorCode, NativeHandle, ProcessId, RefEntry};
+use std::cell::Cell;
 
 fn fixture_pid(fixture: &HostedFixture) -> u32 {
     fixture.process_id()
@@ -108,6 +112,91 @@ fn a_live_token_answers_honestly() {
     assert!(!read.evidence.role.is_unknown());
 }
 
+/// A regression pin for the corpse race A14-9 measured: a token that reads
+/// live at the pre-read check must not be trusted for the rest of the call.
+/// Driven through the fake-corroborator injection seam on
+/// `read_live_element_core` rather than a real process kill, because timing a
+/// real kill against the read without a race is not possible - the fake
+/// answers live on the pre-read call and dead on every call after, the
+/// deterministic stand-in for a process that dies during the read. Removing
+/// the post-read corroboration call `read_live_element_core` makes would turn
+/// this `Err` into an `Ok` built from a real, successfully-read fixture
+/// element.
+#[test]
+fn a_process_that_dies_during_the_read_settles_stale_not_ok() {
+    ensure_test_apartment();
+    let fixture = HostedFixture::spawn().expect("a fixture host starts");
+    let handle = verified_handle(&fixture).expect("a verified handle");
+    let element = handle
+        .downcast_ref::<UIAElement>()
+        .expect("a UIAElement payload");
+    let calls = Cell::new(0u32);
+    let corroborate = |_: &UIAElement| {
+        let already_called = calls.get();
+        calls.set(already_called + 1);
+        if already_called == 0 {
+            Ok(())
+        } else {
+            Err(super::stale_reader_error())
+        }
+    };
+
+    let error = match read_live_element_core(element, deadline(), corroborate, read_live) {
+        Err(error) => error,
+        Ok(_) => panic!("a token that goes dead mid-call must not read Ok"),
+    };
+    assert_eq!(error.code, ErrorCode::StaleRef);
+    assert_eq!(
+        calls.get(),
+        2,
+        "both the pre-read and the post-read corroboration call must run"
+    );
+}
+
+/// The vanished-but-live-process shape Finding B measured: the owning process
+/// answers corroboration honestly, and only the specific resolved element is
+/// gone. A read whose own errors already carry the
+/// `UIA_E_ELEMENTNOTAVAILABLE` disposition - built through the same
+/// `uia_failure_error` classifier a real UI Automation failure would go
+/// through - must settle a complete `STALE_REF`, never fall through to the
+/// completeness gate's retryable `AppUnresponsive`.
+#[test]
+fn a_vanished_resolved_target_settles_stale_not_retryable_unresponsive() {
+    ensure_test_apartment();
+    let fixture = HostedFixture::spawn().expect("a fixture host starts");
+    let handle = verified_handle(&fixture).expect("a verified handle");
+    let element = handle
+        .downcast_ref::<UIAElement>()
+        .expect("a UIAElement payload");
+    let vanished_read = |_: &UIAElement| {
+        let error = uia_failure_error(
+            UiaFailure::Hresult(UIA_E_ELEMENTNOTAVAILABLE),
+            "read an element property",
+        );
+        (ElementProperties::from_reads(Vec::new()), vec![error])
+    };
+
+    let error = match read_live_element_core(
+        element,
+        deadline(),
+        corroborate_verified_process,
+        vanished_read,
+    ) {
+        Err(error) => error,
+        Ok(_) => panic!("a vanished resolved target must not read as complete"),
+    };
+    assert_eq!(error.code, ErrorCode::StaleRef);
+    assert_eq!(
+        error
+            .details
+            .as_ref()
+            .and_then(|details| details.get("complete"))
+            .and_then(serde_json::Value::as_bool),
+        Some(true),
+        "a vanished target settles complete, it is never retried"
+    );
+}
+
 /// The reader-path completeness discipline: an element whose essential slots
 /// are all known passes; the essential-Unknown retryable failure is driven by
 /// the shared read's completeness gate, exercised here through a handle that
@@ -131,10 +220,10 @@ fn the_payload_type_round_trips_through_a_native_handle() {
     assert_eq!(token, "token");
 }
 
+/// Reuses the same construction `resolve_tests` drives: an id-less, name-less
+/// ref with a positive-area bounds hash and an absolute path to the secure
+/// edit, which resolves through the path-and-geometry tier.
 fn blank_secure_entry(fixture: &HostedFixture) -> RefEntry {
-    // Reuses the same construction resolve_tests drives: an id-less, name-less
-    // ref with a positive-area bounds hash and an absolute path to the secure
-    // edit resolves through the path-and-geometry tier.
     let deadline = deadline();
     let root = root_from_hwnd(fixture.handle(), deadline).expect("a fixture root");
     let source = crate::tree::walker_source::UiaTreeSource::for_root(&root).expect("a tree source");
