+++ scripts/check-rust-file-size.sh
@@ -16,8 +16,26 @@ while IFS= read -r file; do
     fi
 done < <(git ls-files --cached --others --exclude-standard -- '*.rs')
 
+# The comment half of this gate is only a gate if it can run. Hardcoding
+# `python3` makes it unrunnable on a Windows developer box, where the
+# interpreter is `python` or `py`: the run then fails for the wrong reason and
+# reads as environment noise rather than a real violation, which is how a
+# comment rule stops being enforced anywhere but CI.
+interpreter=""
+for candidate in python3 python py; do
+    if command -v "$candidate" >/dev/null 2>&1 && "$candidate" -c '' >/dev/null 2>&1; then
+        interpreter="$candidate"
+        break
+    fi
+done
+
+if [ -z "$interpreter" ]; then
+    printf 'no working Python 3 interpreter on PATH (tried python3, python, py)\n' >&2
+    exit 1
+fi
+
 if ! git ls-files -z --cached --others --exclude-standard -- '*.rs' \
-    | python3 scripts/check_rust_comments.py; then
+    | "$interpreter" scripts/check_rust_comments.py; then
     failed=1
 fi
